consolidate format file
