+++ apps/web/src/components/ui/table.tsx
@@ -1,10 +1,9 @@
+import { cn } from 'cnfast';
 import * as React from 'react';
 
 import { Badge, type badgeVariants } from '@/components/ui/badge';
 import { Skeleton } from '@/components/ui/skeleton';
-import { formatUsdCost } from '@/lib/format-cost';
-import { cn } from 'cnfast';
-
+import { formatUsdCost } from '@/lib/format';
 import type { VariantProps } from 'class-variance-authority';
 
 type TableContainerProps = React.HTMLAttributes<HTMLDivElement> & { bordered?: boolean };
