+++ apps/web/src/lib/format.ts
@@ -11,3 +11,16 @@ export function formatTimeAgo(value: Date, now: number = Date.now()): string {
   if (seconds < 86_400) return `${Math.floor(seconds / 3_600)}h ago`;
   return `${Math.floor(seconds / 86_400)}d ago`;
 }
+
+export function formatUsdCost(costUsd: number): string {
+  if (costUsd === 0) {
+    return '$0';
+  }
+
+  if (Math.abs(costUsd) < 0.01) {
+    const precision = Math.min(Math.ceil(-Math.log10(Math.abs(costUsd))) + 1, 8);
+    return `$${costUsd.toFixed(precision).replace(/0+$/, '')}`;
+  }
+
+  return `$${costUsd.toFixed(2)}`;
+}
